Changed expressions to return floats rather than ints, which will make them a bit more useful. They already did their math with floating point precision, just now they don't truncate the result.

+++ Haferlib/src/org/haferlib/util/expression/ConstantExpression.java
@@ -5,14 +5,14 @@ package org.haferlib.util.expression;
 
 public class ConstantExpression implements Expression {
 	
-	private int value;
+	private float value;
 	
-	public ConstantExpression(int value) {
+	public ConstantExpression(float value) {
 		this.value = value;
 	}
 
 	@Override
-	public int getValue() {
+	public float getValue() {
 		return value;
 	}
